Refactor: Implement EmptyStateComponent and improve SettingsScreen
-   Added `EmptyStateComponent` to `QuizScreenComponents.kt` for handling empty states in UI.
-   Moved the `EmptyStateComponent` from `QuizDetailScreen.kt` to `QuizScreenComponents.kt` to make it reusable.
-   Improved `SettingsScreen`:
    -   Updated Card color in SettingsSection.
    -   Removed Divider from SettingsSection.
