+++ app/src/main/java/com/nguyenmoclam/tutorialyoutubemadesimple/ui/screens/SettingScreen.kt
@@ -24,7 +24,6 @@ import androidx.compose.material.icons.filled.Storage
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.Divider
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
@@ -40,6 +39,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
@@ -103,22 +103,6 @@ fun SettingScreen(
             }
 
             Spacer(modifier = Modifier.height(16.dp))
-
-//            // Quiz Configuration Section
-//            SettingsSection(
-//                title = "Quiz Configuration",
-//                icon = Icons.Default.Quiz
-//            ) {
-//                QuizConfigSettings(
-//                    state = state,
-//                    onQuestionOrderChanged = viewModel::setQuestionOrder,
-//                    onMaxRetryCountChanged = viewModel::setMaxRetryCount,
-//                    onShowAnswerAfterWrongChanged = viewModel::setShowAnswerAfterWrong,
-//                    onAutoNextQuestionChanged = viewModel::setAutoNextQuestion
-//                )
-//            }
-//
-//            Spacer(modifier = Modifier.height(16.dp))
 
             // Google Account Section
             SettingsSection(
@@ -328,7 +312,7 @@ fun SettingScreen(
 @Composable
 fun SettingsSection(
     title: String,
-    icon: androidx.compose.ui.graphics.vector.ImageVector,
+    icon: ImageVector,
     content: @Composable () -> Unit
 ) {
     Card(
@@ -336,10 +320,10 @@ fun SettingsSection(
             .fillMaxWidth()
             .padding(vertical = 4.dp),
         colors = CardDefaults.cardColors(
-            containerColor = MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.5f)
+            containerColor = MaterialTheme.colorScheme.surface
         ),
         elevation = CardDefaults.cardElevation(
-            defaultElevation = 2.dp,
+            defaultElevation = 1.dp,
             pressedElevation = 0.dp,
             focusedElevation = 4.dp
         ),
@@ -363,11 +347,6 @@ fun SettingsSection(
                     fontWeight = FontWeight.Bold
                 )
             }
-
-            Divider(
-                color = MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.2f),
-                thickness = 1.dp
-            )
 
             Spacer(modifier = Modifier.height(16.dp))
